# api/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Developer Portal (ADP) API
  version: 1.0.0-alpha
  description: |
    REST API for AI agent governance, audit trails, and documentation.
    Supports both PostgreSQL and SQLite backends.

    ## Overview

    ADP provides two core capabilities for AI coding agents:

    1. **Decision Audit Trail** -- every agent decision is logged with reasoning, confidence, and policy outcome.
    2. **Documentation Engine** -- auto-generates session summaries, risk reports, and pattern analysis from decision records.

    ## Response Formats

    The REST API (`adp-server`, port 8080) uses envelope wrappers for all responses:

    - **Single object**: `{"data": <object>}`
    - **Created object** (201): `{"data": <object>}`
    - **List**: `{"items": [...], "total": 10, "limit": 50, "offset": 0}`
    - **Error**: `{"error": "message", "code": "bad_request", "details": null}`

    ## Sidecar vs REST API

    The commit endpoints (`/v1/commits/*`) exist in two places with **different response formats**:

    - **REST API** (this spec, port 8080): Wraps responses in `{"data": {...}}` envelopes.
    - **MCP HTTP sidecar** (port 8081): Returns top-level JSON without envelopes. Git hooks expect the sidecar format.

    See the commit endpoint descriptions for details on both formats.

    ## Authentication

    ADP supports three authentication methods:

...

    The first registered user is automatically granted the admin role.
...

    ## Database Backends

...

...
servers:
...
paths:
...
  /v1/policies:
    post:
      operationId: createPolicy
      summary: Create a policy definition
      description: |
        Creates a new policy definition. Policies are evaluated in priority order
        during `checkAction` calls. Three policy types are supported:

        - **builtin**: Hardcoded evaluators (sensitive files, blast radius, off-hours, cost limits, migration approval, rate limiting).
        - **rego**: Custom OPA/Rego policies evaluated at runtime.
        - **custom**: Database-stored policies with JSON configuration.
      tags: [Policies]
      security:
        - bearerAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePolicyRequest"
            example:
              name: "no-secrets-in-commits"
              description: "Prevent committing files that may contain secrets"
              category: "security"
              enabled: true
              priority: 100
              policy_type: "builtin"
              builtin_name: "sensitive_files"
              min_trust_level: 1
              tags: ["security", "secrets"]
      responses:
        "201":
          description: Policy created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/PolicyDefinition"
        "400":
          description: Missing required fields (name, category)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Policy with this name already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "policy with name 'no-secrets-in-commits' already exists"
                code: "conflict"
                details: null
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      operationId: listPolicies
      summary: List policy definitions
      description: Returns a paginated list of policy definitions with optional filters.
      tags: [Policies]
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: category
          in: query
          schema:
            type: string
          description: Filter by policy category (e.g., security, compliance, operational)
        - name: policy_type
          in: query
          schema:
            type: string
            enum: [builtin, rego, custom]
          description: Filter by policy type
        - name: enabled
          in: query
          schema:
            type: boolean
          description: Filter by enabled/disabled status
        - name: tags
          in: query
          schema:
            type: string
          description: Filter by tag (single tag value)
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Paginated list of policy definitions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/policies/{id}:
    get:
      operationId: getPolicy
      summary: Get a policy definition by ID
      description: Returns the policy definition including trigger statistics (trigger count, last triggered time).
      tags: [Policies]
      security:
        - bearerAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/PolicyId"
      responses:
        "200":
          description: Policy definition details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/PolicyDefinition"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

    patch:
      operationId: updatePolicy
      summary: Update a policy definition
      description: Partially updates a policy definition. Only provided fields are modified.
      tags: [Policies]
      security:
        - bearerAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/PolicyId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePolicyRequest"
            example:
              priority: 200
              min_trust_level: 3
              tags: ["security", "secrets", "high-priority"]
      responses:
        "200":
          description: Updated policy definition
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/PolicyDefinition"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Conflict (duplicate name)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      operationId: deletePolicy
      summary: Delete a policy definition
      tags: [Policies]
      security:
        - bearerAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/PolicyId"
      responses:
        "204":
          description: Policy deleted (no content)
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/policies/{id}/toggle:
    patch:
      operationId: togglePolicyEnabled
      summary: Toggle a policy's enabled state
      description: |
        Flips the `enabled` flag on a policy definition. If the policy is
        currently enabled it becomes disabled, and vice versa.
      tags: [Policies]
      security:
        - bearerAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/PolicyId"
      responses:
        "200":
          description: Policy with toggled enabled state
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/PolicyDefinition"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------------------------------------------------------------------------
  # Services
  # ---------------------------------------------------------------------------
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT authentication via locally-generated tokens (POST /v1/auth/login)
        or external JWKS provider. Required in production (PostgreSQL mode).
        Optional in SQLite development mode.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Shared API key authentication. Set via ADP_API_KEY environment variable.

  # ---------------------------------------------------------------------------
  # Reusable Parameters
  # ---------------------------------------------------------------------------
  parameters:
...
    PolicyId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Policy definition ID
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 500
      description: Maximum number of items to return
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
      description: Number of items to skip for pagination

  # ---------------------------------------------------------------------------
  # Reusable Responses
  # ---------------------------------------------------------------------------
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "not found"
            code: "not_found"
            details: null
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "authentication required"
            code: "unauthorized"
            details: null
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "internal server error"
            code: "internal_error"
            details: null

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
          enum:
            - bad_request
            - not_found
            - unauthorized
            - forbidden
            - internal_error
            - conflict
        details:
          description: Additional error context (nullable)
          oneOf:
            - type: object
            - type: "null"

    # -------------------------------------------------------------------------
    # Sessions
    # -------------------------------------------------------------------------
...
    CreatePolicyRequest:
      type: object
      required: [name, category]
      properties:
        name:
          type: string
          description: Unique policy name
        description:
          type: string
          description: Human-readable policy description
        category:
          type: string
          description: Policy category (e.g., security, compliance, operational)
        enabled:
          type: boolean
          default: true
          description: Whether the policy is active
        priority:
          type: integer
          default: 0
          description: Evaluation priority (higher = evaluated first)
        policy_type:
          type: string
          enum: [builtin, rego, custom]
          description: Type of policy implementation
        rego_code:
          type: string
          description: OPA/Rego source code (for rego type policies)
        builtin_name:
          type: string
          description: |
            Name of the builtin evaluator (for builtin type policies).
            Available builtins: sensitive_files, blast_radius, off_hours,
            cost_limits, migration_approval, rate_limiting.
        config:
          type: object
          additionalProperties: true
          description: Policy-specific configuration
        min_trust_level:
          type: integer
          minimum: 1
          maximum: 5
          description: Minimum trust level required to bypass this policy
        tags:
          type: array
          items:
            type: string
          description: Tags for filtering and grouping
        metadata:
          type: object
          additionalProperties: true
          description: Arbitrary metadata
    UpdatePolicyRequest:
      type: object
      description: All fields are optional. Only provided fields are updated.
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        enabled:
          type: boolean
        priority:
          type: integer
        policy_type:
          type: string
          enum: [builtin, rego, custom]
        rego_code:
          type: string
        builtin_name:
          type: string
        config:
          type: object
          additionalProperties: true
        min_trust_level:
          type: integer
          minimum: 1
          maximum: 5
        tags:
          type: array
          items:
            type: string
        metadata:
          type: object
          additionalProperties: true
    PolicyDefinition:
      type: object
      properties:
        id:
          type: string
          description: Unique policy definition ID
        name:
          type: string
        description:
          type: string
        category:
          type: string
        enabled:
          type: boolean
        priority:
          type: integer
        policy_type:
          type: string
          enum: [builtin, rego, custom]
        rego_code:
          type: string
        builtin_name:
          type: string
        config:
          type: object
          additionalProperties: true
        min_trust_level:
          type: integer
          minimum: 1
          maximum: 5
        tags:
          type: array
          items:
            type: string
        metadata:
          type: object
          additionalProperties: true
        trigger_count:
          type: integer
          description: Number of times this policy has been triggered
        last_triggered:
          type: string
          description: Relative time since last trigger (e.g., "5 minutes ago", "Never")
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PolicyListResponse:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PolicyDefinition"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer

    # -------------------------------------------------------------------------
    # Services
    # -------------------------------------------------------------------------
...